+++ gradient/api_sdk/clients/model_client.py
@@ -14,3 +14,11 @@ class ModelsClient(BaseClient):
         repository = repositories.ListModels(api_key=self.api_key, logger=self.logger)
         models_list = repository.list(experiment_id=experiment_id, project_id=project_id)
         return models_list
+
+    def delete(self, model_id):
+        """Delete a model
+
+        :param str model_id: Model ID
+        """
+        repository = repositories.DeleteModel(api_key=self.api_key, logger=self.logger)
+        repository.delete(model_id)