+++ gradient/api_sdk/repositories/models.py
@@ -1,6 +1,7 @@
+import gradient.api_sdk.config
 from gradient import config
 from .. import serializers
-from ..repositories.common import ListResources
+from ..repositories.common import ListResources, DeleteResource
 
 
 class ParseModelDictMixin(object):
@@ -15,7 +16,7 @@ class ParseModelDictMixin(object):
 
 class GetBaseModelsApiUrlMixin(object):
     def _get_api_url(self, **_):
-        return config.config.CONFIG_HOST
+        return gradient.api_sdk.config.config.CONFIG_HOST
 
 
 class ListModels(GetBaseModelsApiUrlMixin, ParseModelDictMixin, ListResources):
@@ -45,3 +46,15 @@ class ListModels(GetBaseModelsApiUrlMixin, ParseModelDictMixin, ListResources):
 
         json_ = {"filter": {"where": {"and": [filters]}}}
         return json_
+
+
+class DeleteModel(GetBaseModelsApiUrlMixin, DeleteResource):
+    def get_request_url(self, **kwargs):
+        return "/mlModels/deleteModel/"
+
+    def _send_request(self, client, url, json_data=None):
+        response = client.post(url, json=json_data)
+        return response
+
+    def _get_request_json(self, kwargs):
+        return kwargs