+++ gradient/api_sdk/repositories/tensorboards.py
@@ -1,6 +1,6 @@
 from gradient.api_sdk.serializers import TensorboardSchema
 
-from gradient.config import config
+from gradient.api_sdk.config import config
 from . import common
 from .. import serializers, models
 
@@ -97,7 +97,3 @@ class UpdateTensorboard(ParseTensorboardMixin, GetTensorboardApiUrlMixin, common
 class DeleteTensorboard(GetTensorboardApiUrlMixin, common.DeleteResource):
     def get_request_url(self, **kwargs):
         return "/tensorboards/v1/{}".format(kwargs["id"])
-
-    def _send_request(self, client, url, json_data=None):
-        response = client.delete(url, json=json_data)
-        return response