+++ gradient/api_sdk/serializers/job.py
@@ -1,7 +1,7 @@
 import marshmallow
 
 from gradient.api_sdk import models
-from . import BaseSchema
+from .base import BaseSchema
 
 
 class JobSchema(BaseSchema):
@@ -85,3 +85,4 @@ class JobSchema(BaseSchema):
     rel_dockerfile_path = marshmallow.fields.Str(dump_to="relDockerfilePath", load_from="relDockerfilePath")
     registry_username = marshmallow.fields.Str(dump_to="registryUsername", load_from="registryUsername")
     registry_password = marshmallow.fields.Str(dump_to="registryPassword", load_from="registryPassword")
+    build_only = marshmallow.fields.Bool(dump_to="buildOnly", load_from="buildOnly")