+++ gradient/cli/experiments.py
@@ -1,10 +1,9 @@
-import collections
 import functools
 
 import click
 
-from gradient import constants, utils, logger, workspace
-from gradient.api_sdk.clients import http_client
+from gradient import utils, logger, workspace
+from gradient.api_sdk import constants
 from gradient.cli import common, validators
 from gradient.cli.cli import cli
 from gradient.cli.cli_types import json_string, ChoiceType
@@ -12,14 +11,6 @@ from gradient.cli.common import api_key_option, ClickGroup
 from gradient.cli.utils.flag_with_value import GradientRegisterReaderOption, GradientRegisterWriterOption, \
     GradientRegisterWriterCommand
 from gradient.commands import experiments as experiments_commands
-from gradient.config import config
-
-MULTI_NODE_EXPERIMENT_TYPES_MAP = collections.OrderedDict(
-    (
-        ("GRPC", constants.ExperimentType.GRPC_MULTI_NODE),
-        ("MPI", constants.ExperimentType.MPI_MULTI_NODE),
-    )
-)
 
 MULTI_NODE_CREATE_EXPERIMENT_COMMANDS = {
     constants.ExperimentType.GRPC_MULTI_NODE: experiments_commands.CreateMultiNodeExperimentCommand,
@@ -33,23 +24,22 @@ MULTI_NODE_RUN_EXPERIMENT_COMMANDS = {
 
 
 def get_workspace_handler(api_key):
-    client = http_client.API(config.CONFIG_EXPERIMENTS_HOST, api_key=api_key)
     logger_ = logger.Logger()
-    workspace_handler = workspace.S3WorkspaceHandlerWithProgressbar(experiments_api=client, logger_=logger_)
+    workspace_handler = workspace.S3WorkspaceHandlerWithProgressbar(api_key=api_key, logger_=logger_)
     return workspace_handler
 
 
 @cli.group("experiments", help="Manage experiments", cls=ClickGroup)
-def experiments():
+def experiments_group():
     pass
 
 
-@experiments.group("create", help="Create new experiment", cls=ClickGroup)
+@experiments_group.group("create", help="Create new experiment", cls=ClickGroup)
 def create_experiment():
     pass
 
 
-@experiments.group(name="run", help="Create and start new experiment", cls=ClickGroup)
+@experiments_group.group(name="run", help="Create and start new experiment", cls=ClickGroup)
 def create_and_start_experiment():
     pass
 
@@ -164,7 +154,7 @@ def common_experiment_create_multi_node_options(f):
         click.option(
             "--experimentType",
             "experiment_type_id",
-            type=ChoiceType(MULTI_NODE_EXPERIMENT_TYPES_MAP, case_sensitive=False),
+            type=ChoiceType(constants.MULTI_NODE_EXPERIMENT_TYPES_MAP, case_sensitive=False),
             required=True,
             help="Experiment Type",
             cls=common.GradientOption,
@@ -542,7 +532,7 @@ def create_and_start_single_node(ctx, api_key, show_logs, use_vpc, tensorboard, 
         ctx.invoke(list_logs, experiment_id=experiment_id, line=0, limit=100, follow=True, api_key=api_key)
 
 
-@experiments.command("start", help="Start experiment")
+@experiments_group.command("start", help="Start experiment")
 @click.argument("id", cls=common.GradientArgument)
 @click.option(
     "--logs",
@@ -568,7 +558,7 @@ def start_experiment(ctx, id, show_logs, api_key, options_file, use_vpc):
         ctx.invoke(list_logs, experiment_id=id, line=0, limit=100, follow=True, api_key=api_key)
 
 
-@experiments.command("stop", help="Stop experiment")
+@experiments_group.command("stop", help="Stop experiment")
 @click.argument("id", cls=common.GradientArgument)
 @api_key_option
 @click.option(
@@ -584,7 +574,7 @@ def stop_experiment(id, api_key, options_file, use_vpc):
     command.execute(id, use_vpc=use_vpc)
 
 
-@experiments.command("list", help="List experiments")
+@experiments_group.command("list", help="List experiments")
 @click.option("--projectId", "-p", "project_ids", multiple=True, cls=common.GradientOption)
 @api_key_option
 @common.options_file
@@ -593,7 +583,7 @@ def list_experiments(project_ids, api_key, options_file):
     command.execute(project_id=project_ids)
 
 
-@experiments.command("details", help="Show detail of an experiment")
+@experiments_group.command("details", help="Show detail of an experiment")
 @click.argument("id", cls=common.GradientArgument)
 @api_key_option
 @common.options_file
@@ -602,7 +592,7 @@ def get_experiment_details(id, options_file, api_key):
     command.execute(id)
 
 
-@experiments.command("logs", help="List experiment logs")
+@experiments_group.command("logs", help="List experiment logs")
 @click.option(
     "--experimentId",
     "experiment_id",
@@ -635,3 +625,12 @@ def get_experiment_details(id, options_file, api_key):
 def list_logs(experiment_id, line, limit, follow, options_file, api_key=None):
     command = experiments_commands.ExperimentLogsCommand(api_key=api_key)
     command.execute(experiment_id, line, limit, follow)
+
+
+@experiments_group.command("delete", help="Delete an experiment")
+@click.argument("id", cls=common.GradientArgument)
+@api_key_option
+@common.options_file
+def delete_experiment(id, options_file, api_key):
+    command = experiments_commands.DeleteExperimentCommand(api_key=api_key)
+    command.execute(id)