+++ gradient/commands/deployments.py
@@ -5,9 +5,11 @@ import six
 import terminaltables
 from halo import halo
 
-from gradient import version, logger as gradient_logger, api_sdk, exceptions
+from gradient import version, logger as gradient_logger, exceptions
+from gradient.api_sdk import sdk_exceptions
 from gradient.api_sdk.clients import http_client
-from gradient.config import config
+from gradient.api_sdk.config import config
+from gradient.api_sdk.utils import urljoin
 from gradient.utils import get_terminal_lines
 
 default_headers = {"X-API-Key": config.PAPERSPACE_API_KEY,
@@ -33,6 +35,11 @@ class CreateDeploymentCommand(_DeploymentCommand):
             deployment_id = self.deployment_client.create(use_vpc=use_vpc, **kwargs)
 
         self.logger.log("New deployment created with id: {}".format(deployment_id))
+        self.logger.log(self.get_instance_url(deployment_id))
+
+    def get_instance_url(self, instance_id):
+        url = urljoin(config.WEB_URL, "/console/deployments/{}".format(instance_id))
+        return url
 
 
 class ListDeploymentsCommand(_DeploymentCommand):
@@ -47,7 +54,7 @@ class ListDeploymentsCommand(_DeploymentCommand):
     def _get_instances(self, use_vpc=False, **kwargs):
         try:
             instances = self.deployment_client.list(use_vpc=use_vpc, **kwargs)
-        except api_sdk.GradientSdkError as e:
+        except sdk_exceptions.GradientSdkError as e:
             raise exceptions.ReceivingDataFailedError(e)
 
         return instances
@@ -95,3 +102,9 @@ class StopDeploymentCommand(_DeploymentCommand):
     def execute(self, use_vpc=False, **kwargs):
         self.deployment_client.stop(use_vpc=use_vpc, **kwargs)
         self.logger.log("Deployment stopped")
+
+
+class DeleteDeploymentCommand(_DeploymentCommand):
+    def execute(self, **kwargs):
+        self.deployment_client.delete(**kwargs)
+        self.logger.log("Deployment deleted")