+++ gradient/commands/experiments.py
@@ -6,7 +6,10 @@ import terminaltables
 from click import style
 from halo import halo
 
-from gradient import constants, api_sdk, exceptions, TensorboardClient
+from gradient import api_sdk, exceptions, TensorboardClient
+from gradient.api_sdk import constants, sdk_exceptions
+from gradient.api_sdk.config import config
+from gradient.api_sdk.utils import urljoin
 from gradient.commands import tensorboards as tensorboards_commands
 from gradient.commands.common import BaseCommand, ListCommandMixin
 from gradient.logger import Logger
@@ -84,8 +87,14 @@ class BaseCreateExperimentCommandMixin(object):
             experiment_id = self._create(json_, use_vpc=use_vpc)
 
         self.logger.log(self.CREATE_SUCCESS_MESSAGE_TEMPLATE.format(experiment_id))
+        self.logger.log(self.get_instance_url(experiment_id, json_["project_id"]))
+
         self._maybe_add_to_tensorboard(add_to_tensorboard, experiment_id, self.api_key)
         return experiment_id
+
+    def get_instance_url(self, instance_id, project_id):
+        url = urljoin(config.WEB_URL, "console/projects/{}/experiments/{}".format(project_id, instance_id))
+        return url
 
     def _handle_workspace(self, instance_dict):
         handler = self.workspace_handler.handle(instance_dict)
@@ -125,6 +134,7 @@ class CreateMultiNodeExperimentCommand(BaseCreateExperimentCommandMixin, BaseExp
 
 class CreateMpiMultiNodeExperimentCommand(BaseCreateExperimentCommandMixin, BaseExperimentCommand):
     def _create(self, json_, use_vpc=False):
+        json_.pop("experiment_type_id", None)  # for MPI there is no experiment_type_id parameter in client method
         handle = self.client.create_mpi_multi_node(use_vpc=use_vpc, **json_)
         return handle
 
@@ -143,6 +153,7 @@ class CreateAndStartMpiMultiNodeExperimentCommand(BaseCreateExperimentCommandMix
     CREATE_SUCCESS_MESSAGE_TEMPLATE = "New experiment created and started with ID: {}"
 
     def _create(self, json_, use_vpc=False):
+        json_.pop("experiment_type_id", None)  # for MPI there is no experiment_type_id parameter in client method
         handle = self.client.run_mpi_multi_node(use_vpc=use_vpc, **json_)
         return handle
 
@@ -181,7 +192,7 @@ class ListExperimentsCommand(ListCommandMixin, BaseExperimentCommand):
         project_id = kwargs.get("project_id")
         try:
             instances = self.client.list(project_id)
-        except api_sdk.GradientSdkError as e:
+        except sdk_exceptions.GradientSdkError as e:
             raise exceptions.ReceivingDataFailedError(e)
 
         return instances
@@ -212,7 +223,7 @@ class GetExperimentCommand(BaseExperimentCommand):
         """
         try:
             instance = self.client.get(id_)
-        except api_sdk.GradientSdkError as e:
+        except sdk_exceptions.GradientSdkError as e:
             raise exceptions.ReceivingDataFailedError(e)
 
         return instance
@@ -337,3 +348,9 @@ class ExperimentLogsCommand(BaseExperimentCommand):
         return (style(fg="blue", text=experiment_id),
                 style(fg="red", text=str(log_row.line)),
                 log_row.message)
+
+
+class DeleteExperimentCommand(BaseExperimentCommand):
+    def execute(self, experiment_id, *args, **kwargs):
+        self.client.delete(experiment_id)
+        self.logger.log("Experiment deleted")