+++ tests/functional/test_experiments.py
@@ -1,10 +1,15 @@
+import os
+import tempfile
+import zipfile
+
 import mock
 from click.testing import CliRunner
 
-from gradient import constants
+from gradient.api_sdk import constants
 from gradient.api_sdk.clients import http_client
 from gradient.cli import cli
 from tests import example_responses, MockResponse
+from tests.unit.test_archiver_class import create_test_dir_tree
 
 
 class TestExperimentsCreateSingleNode(object):
@@ -21,6 +26,15 @@ class TestExperimentsCreateSingleNode(object):
         "--machineType", "testType",
         "--command", "testCommand",
         "--workspaceUrl", "some-workspace",
+    ]
+    BASIC_OPTIONS_COMMAND_WITH_LOCAL_WORKSPACE = [
+        "experiments", "create", "singlenode",
+        "--name", "exp1",
+        "--projectId", "testHandle",
+        "--container", "testContainer",
+        "--machineType", "testType",
+        "--command", "testCommand",
+        "--workspace",  # local path added in test
     ]
     FULL_OPTIONS_COMMAND = [
         "experiments", "create", "singlenode",
@@ -264,6 +278,88 @@ class TestExperimentsCreateSingleNode(object):
 
         tensorboard_handler_class.assert_called_once_with("some_key")
         tensorboard_handler.maybe_add_to_tensorboard.assert_called_once_with(True, "sadkfhlskdjh")
+
+    @mock.patch("gradient.workspace.s3_uploader.MultipartEncoderWithProgressbar")
+    @mock.patch("gradient.api_sdk.clients.http_client.requests.post")
+    @mock.patch("gradient.api_sdk.clients.http_client.requests.get")
+    def test_should_zip_and_upload_local_workspace_when_local_path_was_passed_to_workspace_option(
+            self, get_patched, post_patched, multipart_encoder_cls_patched):
+        multipart_encoder_patched = mock.MagicMock()
+        multipart_encoder_content_type = mock.MagicMock()
+        # multipart_encoder_patched.content_type = multipart_encoder_content_type
+        multipart_monitor = mock.MagicMock()
+        multipart_monitor.content_type = multipart_encoder_content_type
+        multipart_encoder_patched.get_monitor.return_value = multipart_monitor
+        multipart_encoder_cls_patched.return_value = multipart_encoder_patched
+
+        headers_for_uploading_to_s3 = self.EXPECTED_HEADERS.copy()
+        headers_for_uploading_to_s3["Content-Type"] = multipart_encoder_content_type
+
+        get_patched.return_value = MockResponse(example_responses.GET_PRESIGNED_URL_FOR_S3_BUCKET_RESPONSE_JSON)
+        post_patched.side_effect = [
+            MockResponse(status_code=204),
+            MockResponse(status_code=201),
+        ]
+
+        workspace_path = create_test_dir_tree()
+        zip_file_name = os.path.basename(workspace_path) + ".zip"
+        command = self.BASIC_OPTIONS_COMMAND_WITH_LOCAL_WORKSPACE[:] + [workspace_path]
+        zip_file_path = os.path.join(tempfile.gettempdir(), zip_file_name)
+        temporary_directory_for_extracted_files = os.path.join(tempfile.gettempdir(), "extracted_files")
+        create_experiment_request_json = self.BASIC_OPTIONS_REQUEST.copy()
+        create_experiment_request_json["workspaceUrl"] = \
+            "s3://ps-projects/" + example_responses.GET_PRESIGNED_URL_FOR_S3_BUCKET_RESPONSE_JSON["data"]["fields"][
+                "key"]
+
+        runner = CliRunner()
+        result = runner.invoke(cli.cli, command)
+
+        # assert self.EXPECTED_STDOUT in result.output, result.exc_info
+
+        with zipfile.ZipFile(zip_file_path) as zip_handler:
+            zip_handler.extractall(temporary_directory_for_extracted_files)
+
+        file1_path = os.path.join(temporary_directory_for_extracted_files, "file1")
+        assert os.path.exists(file1_path)
+        assert os.path.isfile(file1_path)
+        with open(file1_path) as h:
+            assert h.read() == "keton"
+
+        file2_path = os.path.join(temporary_directory_for_extracted_files, "subdir1", "file2")
+        assert os.path.exists(file2_path)
+        assert os.path.isfile(file2_path)
+        with open(file2_path) as h:
+            assert h.read() == "keton"
+
+        get_patched.assert_called_once_with(
+            "https://services.paperspace.io/experiments/v1/workspace/get_presigned_url",
+            headers=self.EXPECTED_HEADERS,
+            json=None,
+            params={"projectHandle": "testHandle", "workspaceName": zip_file_name},
+        )
+
+        post_patched.assert_has_calls(
+            [
+                mock.call(
+                    "https://ps-projects.s3.amazonaws.com/",
+                    json=None,
+                    params=None,
+                    headers=headers_for_uploading_to_s3,
+                    files=None,
+                    data=multipart_monitor,
+                ),
+                mock.call(
+                    self.URL,
+                    json=create_experiment_request_json,
+                    params=None,
+                    headers=self.EXPECTED_HEADERS,
+                    files=None,
+                    data=None,
+                ),
+            ]
+        )
+
+        assert result.exit_code == 0
 
 
 class TestExperimentsCreateMultiNode(object):
@@ -553,6 +649,15 @@ class TestExperimentsCreateAndStartSingleNode(TestExperimentsCreateSingleNode):
         "--command", "testCommand",
         "--workspaceUrl", "some-workspace",
         "--no-logs",
+    ]
+    BASIC_OPTIONS_COMMAND_WITH_LOCAL_WORKSPACE = [
+        "experiments", "run", "singlenode",
+        "--name", "exp1",
+        "--projectId", "testHandle",
+        "--container", "testContainer",
+        "--machineType", "testType",
+        "--command", "testCommand",
+        "--workspace",  # local path added in test
     ]
     FULL_OPTIONS_COMMAND = [
         "experiments", "run", "singlenode",
@@ -1087,6 +1192,95 @@ class TestStopExperiment(object):
                                             params=None)
 
 
+class TestDeleteExperiment(object):
+    URL = "https://services.paperspace.io/experiments/v1/experiments/some-id/"
+    COMMAND = ["experiments", "delete", "some-id"]
+    COMMAND_WITH_OPTIONS_FILE = ["experiments", "delete", "--optionsFile", ]  # path added in test
+    EXPECTED_HEADERS = http_client.default_headers.copy()
+    EXPECTED_HEADERS_WITH_CHANGED_API_KEY = http_client.default_headers.copy()
+    EXPECTED_HEADERS_WITH_CHANGED_API_KEY["X-API-Key"] = "some_key"
+    COMMAND_WITH_API_KEY = ["experiments", "delete", "some-id", "--apiKey", "some_key"]
+    RESPONSE_JSON = {"message": "success"}
+    EXPECTED_STDOUT = "Experiment deleted\n"
+
+    NOT_FOUND_JSON_RESPONSE = {"details": "Experiment not found", "error": "Object not found"}
+    NOT_FOUND_EXPECTED_STDOUT = "Failed to delete resource: Experiment not found\nObject not found\n"
+    INVALID_API_KEY_RESPONSE_JSON = {"details": "Incorrect API Key provided", "error": "Forbidden"}
+    INVALID_API_KEY_STDOUT = "Failed to delete resource: Incorrect API Key provided\nForbidden\n"
+
+    @mock.patch("gradient.api_sdk.clients.http_client.requests.delete")
+    def test_should_send_delete_request_and_print_confirmation(self, delete_patched):
+        delete_patched.return_value = MockResponse(self.RESPONSE_JSON, 204)
+        expected_headers = http_client.default_headers.copy()
+        expected_headers["X-API-Key"] = "some_key"
+
+        runner = CliRunner()
+        result = runner.invoke(cli.cli, self.COMMAND)
+
+        assert result.output == self.EXPECTED_STDOUT, result.exc_info
+        delete_patched.assert_called_once_with(self.URL,
+                                               headers=self.EXPECTED_HEADERS,
+                                               json=None,
+                                               params=None)
+
+    @mock.patch("gradient.api_sdk.clients.http_client.requests.delete")
+    def test_should_send_delete_request_with_changed_api_key_when_api_key_option_was_provided(self, delete_patched):
+        delete_patched.return_value = MockResponse(self.RESPONSE_JSON, 204)
+
+        runner = CliRunner()
+        result = runner.invoke(cli.cli, self.COMMAND_WITH_API_KEY)
+
+        assert result.output == self.EXPECTED_STDOUT
+        delete_patched.assert_called_once_with(self.URL,
+                                               headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
+                                               json=None,
+                                               params=None)
+
+    @mock.patch("gradient.api_sdk.clients.http_client.requests.delete")
+    def test_should_read_options_from_config_file(self, delete_patched, experiments_delete_config_path):
+        delete_patched.return_value = MockResponse(self.RESPONSE_JSON, 204)
+        command = self.COMMAND_WITH_OPTIONS_FILE[:] + [experiments_delete_config_path]
+
+        runner = CliRunner()
+        result = runner.invoke(cli.cli, command)
+
+        assert result.output == self.EXPECTED_STDOUT
+        delete_patched.assert_called_once_with(self.URL,
+                                               headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
+                                               json=None,
+                                               params=None)
+
+    @mock.patch("gradient.api_sdk.clients.http_client.requests.delete")
+    def test_should_print_proper_message_when_experiment_was_not_found(self, delete_patched):
+        delete_patched.return_value = MockResponse(self.NOT_FOUND_JSON_RESPONSE, 404)
+        expected_headers = http_client.default_headers.copy()
+        expected_headers["X-API-Key"] = "some_key"
+
+        runner = CliRunner()
+        result = runner.invoke(cli.cli, self.COMMAND)
+
+        assert result.output == self.NOT_FOUND_EXPECTED_STDOUT, result.exc_info
+        delete_patched.assert_called_once_with(self.URL,
+                                               headers=self.EXPECTED_HEADERS,
+                                               json=None,
+                                               params=None)
+
+    @mock.patch("gradient.api_sdk.clients.http_client.requests.delete")
+    def test_should_send_print_proper_message_when_wrong_api_key_was_used(self, delete_patched):
+        delete_patched.return_value = MockResponse(self.INVALID_API_KEY_RESPONSE_JSON, 403)
+        expected_headers = http_client.default_headers.copy()
+        expected_headers["X-API-Key"] = "some_key"
+
+        runner = CliRunner()
+        result = runner.invoke(cli.cli, self.COMMAND)
+
+        assert result.output == self.INVALID_API_KEY_STDOUT, result.exc_info
+        delete_patched.assert_called_once_with(self.URL,
+                                               headers=self.EXPECTED_HEADERS,
+                                               json=None,
+                                               params=None)
+
+
 class TestExperimentLogs(object):
     URL = "https://logs.paperspace.io/jobs/logs"
     COMMAND = ["experiments", "logs", "--experimentId", "some_id"]
@@ -1125,7 +1319,8 @@ class TestExperimentLogs(object):
         assert result.exit_code == 0
 
     @mock.patch("gradient.api_sdk.clients.http_client.requests.get")
-    def test_should_send_get_request_and_print_all_received_logs_when_logs_command_was_used_with_follow_flag(self, get_patched):
+    def test_should_send_get_request_and_print_all_received_logs_when_logs_command_was_used_with_follow_flag(
+            self, get_patched):
         get_patched.return_value = MockResponse(json_data=example_responses.LIST_OF_LOGS_FOR_EXPERIMENT,
                                                 status_code=200)